refactor(#331): split planning_tasks.py into the planning/ package
The 1,887-line module (13 handlers + 2 bases + 6 helpers) becomes 7
modules keyed by plan-authoring stage — base / framing / review / brief /
propose / merge / refinement — mirroring the #152 cycle/ layout.
planning_tasks.py collapses to the 49-line compatibility shim re-exporting
every name; all 20 top-level defs/classes preserved (AST-verified), zero
logic or prompt-text edits. Helper homing per the issue's design table:
the brief-id/planning-content helpers ride with propose (their only use
sites); the refinement time-budget section rides with refinement.
